Moved __getRawRelationData method to AbstractObject
